assignments/assignment_03/rag_chatbot/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
chainlit
# langchain-core
langchain-community
# langchain-text-splitters
langchain-chroma
langchain
arize-phoenix
langchain-huggingface
# faiss-cpu
unstructured-client
langchain-unstructured
